RESUME_TEAM_DEBUGS/requirements.txt
Flask==3.0.0
PyMuPDF
pdfplumber
python-docx


